refactor(notifications): colocate mark-all-read tests with route
Move POST tests to mark-all-read/__tests__/route.test.ts to satisfy
the test coverage checker's colocated test expectation.
